2_airflow/requirements.txt
kaggle
apache-airflow-providers-google
pyarrow
pandas
gcloud